Add solution for Euler problem no. 13
