+++ src/test/java/javaslang/collection/euler/Utils.java
@@ -5,7 +5,13 @@
  */
 package javaslang.collection.euler;
 
+import javaslang.collection.Iterator;
 import javaslang.collection.Stream;
+
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.net.URL;
+import java.util.Scanner;
 
 public final class Utils {
 
@@ -14,6 +20,34 @@ public final class Utils {
 
     public static Stream<Integer> fibonacci() {
         return Stream.of(1, 1).appendSelf(self -> self.zip(self.tail()).map(t -> t._1 + t._2));
+    }
+
+    public static Stream<String> readLines(File file) {
+        final Scanner scanner;
+        try {
+            scanner = new Scanner(file);
+        } catch (FileNotFoundException e) {
+            return Stream.empty();
+        }
+        return Stream.ofAll(new Iterator<String>() {
+            @Override
+            public boolean hasNext() {
+                return scanner.hasNextLine();
+            }
+
+            @Override
+            public String next() {
+                return scanner.nextLine();
+            }
+        });
+    }
+
+    public static File file(String fileName) {
+        final URL resource = Utils.class.getClassLoader().getResource(fileName);
+        if(resource == null) {
+            throw new RuntimeException("resource not found");
+        }
+        return new File(resource.getFile());
     }
 
     public static String reverse(String s) {
